backend/requirements.txt
fastapi
uvicorn
pdfplumber
pandas
typing
pydantic
openpyxl
groq
markdown
python-dotenv
python-multipart